% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create.shell}
\alias{create.shell}
\title{Create a template shell script of mMARCH.AC}
\usage{
create.shell()
}
\value{
The function will create a template shell script of mMARCH.AC in the current directory, names as STUDYNAME_part0.maincall.R
}
\description{
Create a template shell script of mMARCH.AC, named as STUDYNAME_part0.maincall.R.
}
