% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IV2.R
\name{IV2}
\alias{IV2}
\title{Intradaily Variability}
\usage{
IV2(x)
}
\arguments{
\item{x}{\code{vector} of activity data}
}
\value{
IV
}
\description{
This function calcualte intradaily variability, a nonparametric metric
reprsenting fragmentation of circadian rhtymicity
}
\references{
Junrui Di et al. Joint and individual representation of domains of physical activity, sleep, and circadian rhythmicity. Statistics in Biosciences.
}
