% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time2.R
\name{PAfun}
\alias{PAfun}
\title{Timne Metrics for Whole Dataset}
\usage{
PAfun(count.data, weartime, PA.threshold = c(50, 100, 400))
}
\arguments{
\item{count.data}{\code{data.frame} of dimension n*1442 containing the 1440 minute activity data for all n subject days.
The first two columns have to be ID and Day.}

\item{weartime}{\code{data.frame} with dimension of \code{count.data}.
The first two columns have to be ID and Day.}

\item{PA.threshold}{threshold to calculate the time in minutes of sedentary, light, moderate and vigorous activity the data.}
}
\value{
A dataframe with some of the following columns
\item{ID}{identifier of the person}
\item{Day}{indicator of which day of activity it is, can be a numeric vector of sequence 1,2,... or a string of date}
\item{time}{time of certain state}
}
\description{
This function is a whole dataset wrapper for \code{Time}
}
