% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time2.R
\name{Time2}
\alias{Time2}
\title{Time of A Certain activity State}
\usage{
Time2(x, w, thresh, smallerthan = TRUE, bout.length = 1)
}
\arguments{
\item{x}{\code{vector} of activity data.}

\item{w}{\code{vector} of wear flag data with same dimension as \code{x}.}

\item{thresh}{threshold to binarize the data.}

\item{smallerthan}{Find a state that is smaller than a threshold, or greater than or equal to.}

\item{bout.length}{minimum duration of defining an active bout; defaults to 1.}
}
\value{
Time
}
\description{
Calculate the total time of being in certain state, e.g. sedentary, active, MVPA, etc.
}
